refactor: restructure view code CSS
