--- app/examples/user-profile/user-profile-example.xml
@@ -1,68 +1,0 @@
-<examples:ExamplePage
-        cssFile="~/examples/user-profile/user-profile-example.css"
-        xmlns="http://www.nativescript.org/tns.xsd"
-        xmlns:examples="examples/example-base-page"
-        navigatingTo="onPageNavigatingTo"
-        tap="onPageTapped">
-    <Page.actionBar>
-        <ActionBar title="User profile" automationText="ActionBar">
-            <NavigationButton icon="res://ic_back" tap="goBack" automationText="GoBack" />
-            <Android>
-                <ActionItem id="exampleMenuButton" icon="res://ic_menu_example" automationText="ExampleMenu" />
-            </Android>
-            <iOS>
-                <ActionItem id="exampleMenuButton" ios.position="right" automationText="ExampleMenu">
-                    <ActionItem.actionView>
-                        <Image src="res://ic_menu_example" width="22" height="22" margin="0, -11, 0, 11" />
-                    </ActionItem.actionView>
-                </ActionItem>
-            </iOS>
-        </ActionBar>
-    </Page.actionBar>
-    <GridLayout id="example-background" loaded="onBackgroundLoaded">
-        <GridLayout id="form-content" class="form-content" rows="auto,*,auto" loaded="onContentLoaded">
-            <StackLayout class="top-border" row="0"/>
-            <ScrollView row="1" class="fields-section">
-                <StackLayout>
-                    <!-- Username -->
-                    <GridLayout ios:columns="auto,*" android:rows="auto, auto" verticalAlignment="top">
-                        <Label text="Username" class="field-title" />
-                        <TextField id="username" text="{{ username }}" class="field" ios:col="1" android:row="1" tap="onTextInputTapped" />
-                    </GridLayout>
-                    <StackLayout class="line"/>
-                    <!-- Email -->
-                    <GridLayout ios:columns="auto,*" android:rows="auto, auto" verticalAlignment="top">
-                        <Label text="Email" class="field-title" />
-                        <TextField id="email" text="{{ email }}" class="field" ios:col="1" android:row="1" tap="onTextInputTapped" />
-                    </GridLayout>
-                    <StackLayout class="line"/>
-                    <!-- Password -->
-                    <GridLayout ios:columns="auto,*" android:rows="auto, auto" verticalAlignment="top">
-                        <Label text="Password" class="field-title" />
-                        <TextField id="password" text="{{ password }}" secure="{{ showPassword, !showPassword }}" class="field"  ios:col="1" android:row="1" tap="onTextInputTapped" />
-                    </GridLayout>
-                    <StackLayout class="line"/>
-                    <StackLayout orientation="horizontal" tap="onShowPasswordTapped">
-                        <Image src="{{ showPassword ? 'res://checkbox_checked' : 'res://checkbox_unchecked' }}" class="checkbox" />
-                        <Label text="Show password" class="field-switch-title" />
-                    </StackLayout>
-                    <StackLayout class="line"/>
-                    <!-- Bio -->
-                    <StackLayout>
-                        <Label text="Bio" class="field-title" ios:visibility="collapse"/>
-                        <TextView id="bio" text="{{ bio }}" android:hint="Add bio" ios:hint="Bio" class="field-multiline" tap="onTextInputTapped"/>
-                    </StackLayout>
-                    <StackLayout class="line"/>
-                    <!-- Public -->
-                    <GridLayout columns="*, auto" verticalAlignment="top">
-                        <Label text="Public profile" class="field-switch-title" />
-                        <Switch checked="{{ isPublic }}" class="field-switch" />
-                    </GridLayout>
-                </StackLayout>
-            </ScrollView>
-            <Button text="Update" class="update" tap="onUpdateButtonTapped" row="2"/>
-        </GridLayout>
-        <!-- Picture -->
-        <Image id="image" src="res://user" class="profile-picture" tap="onProfilePictureTapped"/>
-    </GridLayout>
-</examples:ExamplePage>
